Clean Ups

+++ code2inv/prog_generator/fuzz/src/120.c
@@ -80,7 +80,7 @@ int main()
   for (;;)
   {
     size_t len;
-    const int32_t *buf;
+    const int16_t *buf;
 
     HF_ITER(&buf, &len);
