+++ code2inv/prog_generator/fuzz/src/70.c
@@ -84,7 +84,7 @@ int main()
   for (;;)
   {
     size_t len;
-    const int32_t *buf;
+    const int16_t *buf;
 
     HF_ITER(&buf, &len);
